first pass at multi-select dblclick
let me know if this interface seems right. i don't like monkey-patching
`grid.on('dblclick', 'tbody tr', /*...*/)`, but it seems better than
some special purpose `grid.onDblClick` or some crap.  like is there a
more robust way to detect a delegate selector like 'tbody tr'?

i think the best way to test it is to run the unit test and just mess
around w/ the grid at the very bottom b/c that makes a document that's
bigger than the window, which does a better job of exercising the math
that matches the dblclick position to a row.