defect 7387 delete filename on show
It may be better to do this on each implementation of a fileupload.
For example on when used on a form we could delete the filename on
successful form submit or form cancel.
Currently that's the comon use case so I'm encapsulating that change
within the fileupload view so we don't have to add that boilerplate
in every form that used this view.